tests/submission/Apple/dataset/dataset-instanceof.html

<!DOCTYPE html>
<html>
  <head>
    <title>Dataset - element.dataset is an instance of DOMStringMap</title>
    <script src="../../../resources/testharness.js"></script>
    <script src="../../../resources/testharnessreport.js"></script>
  </head>
  <body>
    <h1>Dataset - element.dataset is an instance of DOMStringMap</h1>
    <div id="log"></div>
    <script>
      test(function() {
        var d = document.createElement("div");
        assert_true(d.dataset instanceof window.DOMStringMap);
      });
    </script>
  </body>
</html>
